LayoutTests/inspector/debugger/script-formatter-expected.txt

Tests the script formatting functionality.

Debugger was enabled.

Running: testScriptFormatterWorker

Running: testSourceMapping
Debugger was disabled.
